+++ packages/core/src/ceramic.ts
@@ -469,6 +469,7 @@ export class Ceramic implements CeramicApi {
       if (
         (networkOptions.name == Networks.MAINNET || networkOptions.name == Networks.ELP) &&
         anchorServiceUrl !== 'https://cas-internal.3boxlabs.com' &&
+        anchorServiceUrl !== 'https://cas-direct.3boxlabs.com' &&
         anchorServiceUrl !== DEFAULT_ANCHOR_SERVICE_URLS[networkOptions.name]
       ) {
         throw new Error('Cannot use custom anchor service on Ceramic mainnet')
@@ -738,12 +739,10 @@ export class Ceramic implements CeramicApi {
       throw new Error('Writes to streams are not supported in gateway mode')
     }
 
+    const id = normalizeStreamID(streamId)
+    this._logger.verbose(`Apply commit to stream ${id.toString()}`)
     opts = { ...DEFAULT_APPLY_COMMIT_OPTS, ...opts, ...this._loadOptsOverride }
-    const state$ = await this.repository.applyCommit(
-      normalizeStreamID(streamId),
-      commit,
-      opts as CreateOpts
-    )
+    const state$ = await this.repository.applyCommit(id, commit, opts as CreateOpts)
 
     const stream = streamFromState<T>(
       this.context,
@@ -753,6 +752,7 @@ export class Ceramic implements CeramicApi {
     )
 
     await this.repository.indexStreamIfNeeded(state$)
+    this._logger.verbose(`Applied commit to stream ${id.toString()}`)
 
     return stream
   }
@@ -785,6 +785,9 @@ export class Ceramic implements CeramicApi {
     opts = { ...DEFAULT_CREATE_FROM_GENESIS_OPTS, ...opts, ...this._loadOptsOverride }
     const genesisCid = await this.dispatcher.storeCommit(genesis)
     const streamId = new StreamID(type, genesisCid)
+    this._logger.verbose(
+      `Created stream from genesis, StreamID: ${streamId.toString()}, genesis CID: ${genesisCid.toString()}`
+    )
     const state$ = await this.repository.applyCreateOpts(streamId, opts)
     const stream = streamFromState<T>(
       this.context,
@@ -792,6 +795,7 @@ export class Ceramic implements CeramicApi {
       state$.value,
       this.repository.updates$
     )
+    this._logger.verbose(`Created stream ${streamId.toString()} from state`)
 
     await this.repository.indexStreamIfNeeded(state$)