+++ packages/anchor-listener/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-listener@1.15.0-rc.1...@ceramicnetwork/anchor-listener@1.15.0) (2023-07-31)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-listener
+
+
+
+
+
+# [1.15.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-listener@1.15.0-rc.0...@ceramicnetwork/anchor-listener@1.15.0-rc.1) (2023-07-24)
+
+**Note:** Version bump only for package @ceramicnetwork/anchor-listener
+
+
+
+
+
 # [1.15.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/anchor-listener@1.14.0...@ceramicnetwork/anchor-listener@1.15.0-rc.0) (2023-07-13)
 
 **Note:** Version bump only for package @ceramicnetwork/anchor-listener